EventVenueProject/requirements.txt
alembic==1.7.3
aniso8601==9.0.1
blinker==1.4
click==7.1.2
Deprecated==1.2.13
Flask==1.1.4
flask-buzz==0.1.15
Flask-Classful==0.14.2
Flask-Cors==3.0.10
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
flask-praetorian==1.2.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
importlib-resources==5.2.2
inflection==0.3.1
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.5
MarkupSafe==2.0.1
passlib==1.7.4
pendulum==2.1.2
psycopg2-binary==2.9.1
py-buzz==1.0.3
PyJWT==2.1.0
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
pytzdata==2020.1
six==1.16.0
SQLAlchemy==1.4.25
uuid==1.30
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.6.0
